Reorganized files
Since the number of source files is growing, a bit of organization of was made to better sort the different source files. For now, all the files that concern the world generation and / or interaction with it were placed in a world folder
