+++ src/main.c
@@ -12,10 +12,10 @@
 #include "shader.h"
 #include "camera.h"
 #include "material.h"
-#include "blocks.h"
-#include "chunk.h"
-#include "pointer.h"
-#include "world.h"
+#include "world/blocks.h"
+#include "world/chunk.h"
+#include "world/pointer.h"
+#include "world/world.h"
 #include "game_config.h"
 #include "input_process.h"
